Move delete mapping logic to its row component
