+++ frontend/src/metabase/admin/settings/components/widgets/GroupMappingsWidget/DeleteGroupMappingModal/index.tsx
@@ -1,2 +1,2 @@
 export { default } from "./DeleteGroupMappingModal";
-export type { ValueType } from "./DeleteGroupMappingModal";
+export type { DeleteMappingModalValueType } from "./DeleteGroupMappingModal";
